// clients/javascript/lib/gen_types/CalendarEventResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CalendarEventDTO } from './CalendarEventDTO'

/**
 * Calendar event response object
 */
export type CalendarEventResponse = {
  /**
   * Calendar event retrieved
   */
  event: CalendarEventDTO
}